Added a simple test file for readline function
The readline.c file right now just takes input and immediatly prints it
out again. This way it acts as an echo/cat reading from the stdin. If
the line is not NULL and is not the empty string '""', it is added to
the history.

Executables using readline should be compiled with the -lreadline flag.
This is done in the very basic Makefile.

A .gitignore file sets rules so that the executables won't be
accindetally pushed to the repo.

Since readline has memory leaks we don't have to fix, valgrind will
intermix all the errors making it difficult to know if the leak is
something fixable inside our own code or not. I've included a
suppression file that was automatically generated by valgrind.