Added basic exit_handler and add_history
Call exit_handler when you want to exit the minishell executable. Later,
we might need to add an argument holding all the things that need to be
freed before exiting. Currently takes an 'int status' to exit the
program with and clears the history.

In main, only add to the history if there was a line and it is not the
empty string. If user inputs and EOF with 'ctrl-d' line will be returned
as NULL. We exit in that case. To distinguish between EOF and an actual
error, errno is passed.