AST structure does work, printing and freeing
This proves that it works and can be written according to norm rules.
Right now I really dislike how to allocate the full tree as I have to do
it manually. Carefully making sure all the types are correct and such is
a pain.

Next up: a testing program to tokenize a string of math instructions.