Added a dirty linked list to use as input for AST
It is dirty and in minishell would like to write it using the libft
utils.

The purpose for this linked list is to be passed along as argument to
build the AST. It will be the intermediate form between the raw input
from the user and the AST. It will have to be constructed by the
tokenizer, assigning the different values.

I believe we need the depth to know which nodes are next
in queue to be allocated and pointed to. Notice there can only be one
head which is dept 0. After this, there can only two of each depth level
as it corresponds to the left and right branching pointers in the AST. A
branch always ends in an AST_NBR.

If we don't use the depth, I think we'll need a *prev pointer to walk
back.

Other than that, we'll have the tag and values (here just nbr) to know
what we actually have to allocate.

Constructing this will be a whole other beast to tackle. Maybe this
isn't even a great idea :skull: