A pretty simple tokenizer for arithmatic
Making this made me realize, continuing with this would be implementing
something that isn't needed for minishell. The grammar is too different
and I should instead look to implement the rules from bash language.
This is nonetheless pretty valuable to look at for inspiration.

Function 'next_token' allocates the next token using a static size_t to
remember where in the string we were. Ignore whitespace, then look at
one character at a time. Each time joining it to our returning variable.
Maybe more efficient to first count how much should be allocated and
strdup from the input string. Decide on the rules if the token is
finished and return it. Here it's either mathing parenthesis, new
whitespace or the end of the string.

Notice that for the implementation for the AST using a depth, one still
has to parse this list of tokens to assign this value. Or one could make
Parsing a step of building the AST.