Bit of progress translating pipex
Made it possible to create a linked list of tokens (manually coding it
right now, as this doesn't use the tokenizer). I think I might be able
to include the parsing step in the step where we build the AST. This
means there would be no need to do a pass over the tokens and assign a
depth value to it.