Now token list gets divided into the piped parts
With the change of plans to not use an AST, I'll need to look at the
linked list of tokens that comes from the tokenizer and construct a
linked list of commands. Each command will have their arguments and a
linked list of redirects. How many times do you think I can repeat this
linked list of "linked list of" inside this commit message.

What I have done for this commit is construct an array with each element
being a node in the token list. The pipe nodes will have been freed and
the node pointing to said pipe set to point to NULL. This way we have an
array of linked list. Each linked list represents what was typed between
two pipes (or beginning/end of string).

I can already think of a couple of ways to clean this function up. Like
first only put the list heads in the array, and then as a second step
free the pipes for each of the array elements.

Next step is to then go over this array to make our list of commands.