Cleanup from ast stuff, split make_pipe_part_list
The structures not needed anymore because we won't be using an AST have
been removed. This also includes the depth member of the tokens.

I've split up the function to make the array of linked list elements.
Now it just puts the head for each list in the array, to later go
through all of them to free the pipe token and point the previous one to
NULL.