First working version conversing tokens to cmds
I find my code very sloppy here, but It seems to be working. In comes a
linked list of tokens, that have had a bit of parsing giving tags to
them. And out comes a linked list of commands, each with their argument
vector and potentially the linked list of redirects.

There is a lot of things I skipped over since this is just a test. Error
handling and freeing the allocated stuff are the biggest things.

I do think it would now be easier to implement this into pipex, because
pipex already used just a linked list of commands. Main thing that is
different is the redirects.