First working pipex with the new cmd node struct
file minipipe.c is the combination of the files pipex.c and cmds.c. From
a (manual) linked list of tokens, the cmds are made and passed to the
pipex part. There the commands are executed one by one, piping their
output to the next one. Currently ignoring redirects. Lots of
filedescriptors and error handling that still need to be done.

The execve is replaced by execvp to quickly test, but later it should be
replaced back again and first searched for cmd's path.