Added redirects '<', '>' and '>>' when executing
Opens the files with the required flags and redirects the
filedescriptors. Again, freeing stuff and error handling should be
looked at.

This also does not include the here doc as I'm unsure if this is the
best place to implement it. Maybe reading it should happen in one of the
previous steps.

The files are opened with O_CLOEXEC so they close when execve would be
called. I'm probably missing some open filedescriptors though.