Implemented functions to free before exiting.
They are not called yet when errors occur, just made and ran before exit
of the main.

I find it to not be as clean as I want. Since the strings from the token
list have to be malloced (here just strdup), we have to free them. But
these are shared in the pipe_parts as well als later on in the cmd_av
and redirs list. This makes it a bit difficult to just implement the
functions for each as it would mean invalid frees. I'm hoping to rework
it so this is way more manageable.