Added basic pipex part to the project
Make commands from the token list, updated the free_list function to
free_tokens to do just what it name suggests. Only thing that needs to
be freed from the commands then are the nodes of the linked list (also
the nodes from redirects list) as the strings are just pointers to the
elements in the token list.

Able to run and pipe commands (no here_doc yet).

Still a lot to be done. exits_handlers should be considered to instead
of exiting the minishell process, print an error and continue (but still
free the resources). Good example is commands exiting with error,
minishell should still run and just set a variable instead of exiting.