Added env variable stuct and basic functions.
Add to the back, search for a name, delete a target node. The variables
are in a double linked list. This could be usefull later on, or we can
remove the prev pointer and make it single linked.

Want to rename the struct so it is less characters.

Changes in the main are just to test it, nothing more.