Cleaned up use of envp (and some norm stuff)
Now envp is only used to load and make the linked list for environment
variables in the t_minishell structure. Places were envp was previously,
have been changed into a pointer to this t_minishell structure.

Also did a tiny bit of norminette, renaming the global with a g_ in
front, and checking alignments. Removed some comments.