Changed execvp to execve
Had to implement quite a lot to make this change. Needed to search in
the PATH variable in the environment and try paths for executable
access.

execve expects environment variables as a null terminated char ** (like
the one we get from main). Needed to make functions to convert our
linked list of variables to this form as users may add change
or remove variables.

For now this is done only once, at the start, and
stored as a member in the t_minishell structure. This means the
export_env is not made for each command that's ran. Update this whenever
a variable would be exported. Also, if we want to do it like bash with
'local variables' per command, pipex part will have to make a copy and
add any variables the user put in front of the command.

Again skipped a bit of error handling.