Added simple heredoc
Had to change the redir structure to also hold an int fd. Reading the
input from the user is done before executing any commands. It is done by
calling readline with a different prompt message. The line read is
written to a pipe. After user is done (ctrl+d or exact_match delimiter),
the fd of the pipe's read end is stored in the redirect node's
structure. Then later when actually executing the commands, a redirect
is done from this fd.

This matches the behaviour of bash, as it reads everything before
executing. Only thing needed still is variable expansion, and the option
to turn it of by quoting the delimiter.