Potential for correct behaviour with signals
I tried a lot of things here, maybe this is more like the way we want
it, maybe it's not. I don't know yet, and will need to run further
testing.

In the cmd_execute, the wait is now only done for the very last cmd,
meaning pipes are read from as soon as new bytes are written by the
previous command. This means no limit for the pipe buffers are a
non-issue and is the more correct solution. This might end up messing
with some other things in the program though. One thing at least is
writing to the stdout from a child seems to have weird behaviour when
the amount exceeds pipe buffer max.

In the custom_sigint_handler_function the SIGINT is sent to the cmd_pid,
but this would mean only the last cmd would receive it, right?
Currently, after sending the signal, a wait is done and some checks on
the termination status are performed.