Handle SIGCHLD to catch killing and dumping
Create all childs for every cmd and wait for all children at the end.
SIGCHLD gets siginfo_t *info as second argument to know in what state
the child exited or got killed (from SIGINT) or dumped (from SIGQUIT).

Currently the only problem is the "need" for the sleep in ft_wait()
because the color of the arrow would not reflect the state of the
prev_exit correctly. This is because the program moves to the next
prompt before the sig handling is done. Would like to find a way to not
have to do this. Idealy only prompt after handling every signal.