exchanged SIGCHLD handling for wait funcs
Read online that unblocked SIGCHLD handling during a wait is not the
best way to go about it here. I replaced it by waiting on the children
instead and if the wait is interrupted (by SIGINT or SIGQUIT) we try and
get the info from the last child in the command pipeline.

This now handles exactly like bash (I think) which SIGCHLD tried to
implement. The color of the status arrow is now correct as it can't
issue the next prompt before a signal is handled.

Next up is heredoc and its signal interactions. Probably fixed by
reading the input from within a child process, we can reuse our
implementation of this commit for that as well.