Env variable name and value split malloc
Previously the value of the variable was stored in the allocated address
space for the name with a terminating zero where the equal sign in the
token used to be. This was cool in theory but made some other things
annoying to implement.

This commit makes changes so the name and value are two different
allocated strings. This means the value can be more easily set to NULL.
Now this made it pretty easy to make two functions to make variables.
There is the function env_add_var taking in a full token
(of the form "VAR=value") and the function env_add_var2 taking the
name and value as arguments.

This also better mimics the behaviour that bash has. Export a new
variable without a value in the token (thus set to NULL). The variable
is shown in the printing by 'export' but is not included in the exported
environment which the print done by 'env' shows us.

Changes still to come: The variables to be initialized at shell launch
to use the new functions.
There is to my knowledge no need for the function 'env_add_var_only' so
it can be removed.