A small test to see if we want tokens this way
This is just a test and is will probably not be implemented.

First step just splits the input line by whitespace but ignores
whitespace when a quote was encountered. Problem is that in some cases a
token should be delimited when a different char was found and not just
on whitespace. example the pipe char '|' delimits the token before it.
This might be possible to be split in a separate step but seems kinda
stupid. Maybe fixed with a couple of rules to follow?

Quote characters are always included in the token, so expansion step can
remove them.