Heredoc now read from a child process
This allows for a bit of an easier time with the signal handling. SIGINT
ends reading the heredoc and does not continue with the rest of the
inputted line. A warning is printed to stderr if the doc was ended with
an eof instead of the delimiter. This warning message does not include
the line nr. it occured on which bash does give.