A different way of tokenizing and expanding
In this solution we traverse the inputted line character per character
joining them to the token. The token is delimited by certain characters,
which are ignored if we currently have an open quote tag. When we get to
a '$' character, we traverse according to variable naming rules and
don't append the characters. Instead, we search for the variable in the
environment and append its value if it exists and has one.

This solution is great only needing to traverse de line once and no need
to loop multiple times. Potential problems still are: quotes within
quotes, a token with a string NULL is set in certain cases ('echo ' the
extra space starts a new token but never ads characters).

Next step is to see if this integrates well with running the commands as
this was temporarly disabled in previous commit.